add server models to app

+++ app/core/models/models-api/src/commonMain/kotlin/com/diva/models/api/pagination/response/PaginationResponse.kt
@@ -10,7 +10,7 @@ data class PaginationInfo(
     @SerialName("limit")
     val limit: Int,
     @SerialName("total_items")
-    val totalItems: Int,
+    val totalItems: Long,
     @SerialName("total_pages")
     val totalPages: Int,
 )
